updates updates updates

+++ post_process/post_process_undo.sh
@@ -3,6 +3,20 @@ toswitch=$1
 switch_input
 tolistpath=$limitdir
 post_process_dir
+
+if [ ! -z $2 ]
+then
+    if [[ $2 != "combined" ]] && [[ $2 != "joined" ]]
+    then
+        print "Error 2nd arguement is no good"
+        exit
+    else
+        method=$2
+    fi
+else
+    method="combined"
+fi
+
 for tdir in $prepdirs
 do
     echo "CLEANING: ${tdir}"
@@ -10,8 +24,12 @@ do
     #Remove the post processed files
     eval "rm -f ${tdir}/*joined.Rdata"
     eval "rm -rf ${tdir}/joined/"
-    eval "rm -rf ${tdir}/*combined.Rdata"
-    eval "rm -rf ${tdir}/graphs/"
+
+    if [[ $method == "combined" ]]
+    then
+        eval "rm -rf ${tdir}/*combined.Rdata"
+        eval "rm -rf ${tdir}/graphs/"
+    fi
 
     #Move original data back to base directory
     eval "mv ${tdir}/med/*med.Rdata" "${tdir}/."
